boss vinh update

+++ app/src/main/java/com/finger/hsd/model/ProductType_v.kt
@@ -12,7 +12,7 @@ open class ProductType_v : RealmObject(), Serializable {
     var image: String? = null
     var check_product : Boolean = false // Kiểm duyệt sản phẩm
     var check_barcode : Boolean = false // kiểm tra có phải là barcode thật hay không. Có một số sản phẩm không có barcode.
-    var create_at: Int? = null
+    var create_at: Long? = null
     var v: Int? = null
